backend/requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
python-dotenv>=1.0
Pillow>=10.0
djangorestframework~=3.16.0
drf-yasg>=1.20.0
django-redis>=5.4,<6.0
watchdog~=6.0.0
minio~=7.2.15